Test creation of directory and file
