+++ pkg/architest/file.go
@@ -5,22 +5,7 @@ import (
 	"go/parser"
 	"go/token"
 	"path/filepath"
-	"regexp"
 )
-
-var (
-	AcceptanceTestFileRegex *regexp.Regexp
-	AcceptanceTestNameRegex *regexp.Regexp
-	TestFileRegex           *regexp.Regexp
-	TestNameRegex           *regexp.Regexp
-)
-
-func init() {
-	AcceptanceTestFileRegex = regexp.MustCompile("^.*_acceptance_test.go$")
-	AcceptanceTestNameRegex = regexp.MustCompile("^TestAcc_.+$")
-	TestFileRegex = regexp.MustCompile("^.*_test.go$")
-	TestNameRegex = regexp.MustCompile("^Test.*$")
-}
 
 type File struct {
 	packageName string
